Kubrick Mode = true
Added in the beginning of the Camera system.
Check out Purgatory.unity to see it in action. Play around with the CameraSystem script on the CameraSystem object

The system will assign the Main Camera to pre-placed locations around the game world
These positions are assigned with Camera Positions prefabs
It needs at least one "playerCam" tagged object node, preferably parented to the player
static, in-world camera positions should be tagged with "worldCam"