Alright, youhs guise. I done went and made the Nirvana Cubes work.
How do they work, you ask? GOOD QUESTION!
Simply press enter when the player object's new INTERACTION TRIGGER is colliding with the Nirvana Cube, and you will take control of the NIRVANA GRAVITY VOLUME. Pressing left or right while inside the Nirvana Cube will rotate the gravity volume--and, so long as the world camera position is parented to the gravity volume, it will rotate as well. Wouldja look at that!
Press enter again to leave the Nirvana Cube. This will put the player in front of the cube (in the -z direction).
***The level itself does not actually move, but the rotation of the gravity volume and camera make it seem as though it does.***
NOTES FOR NIRVANA LEVEL DESIGN:
The Nirvana Gravity Volume rotates around the z-axis and has been pre-parented to a camera position (remember to add a Camera System to the scene so it works!). The level should be 2.5D, with two layers. All Nirvana cubes should be in the back layer. DO NOT rotate the Nirvana Cubes or Gravity Volume from the prefabs. Any obstacles that should apply to both the player AND the Nirvana Cubes should be present in BOTH layers. Any obstacles that should restrict ONLY the movement of the Nirvana Cubes should be present in the back layer, only. The player can traverse both layers. DO NOT place any obstacles on only the front layer, or else the player might get stuck in them when trying to leave a Nirvana Cube.
Level design for Nirvana should be entirely possible now. If you want to make any objects that are NOT Nirvana Cubes that are still affected by gravity, I also added a BasicGravBody prefab which can be rescaled as desired. Anything other obstacles should have rigidbodies and be kinematic.